requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.12.0
psycopg2-binary==2.9.9
chatterbot==1.0.4
chatterbot-corpus==1.2.0
PyYAML==5.4.1 